+++ sdk/net/basic/tcp_server_socket.hpp
@@ -50,14 +50,11 @@ class tcp_server_socket : public server_socket {
      * @param   char*           ptr_data        [OUT]
      * @param   size_t          size_data       [IN]
      * @param   size_t*         cbread          [OUT]
-     * @param   struct sockaddr* addr           [outopt]
-     * @param   socklen_t* addrlen              [inopt]
-     * @return  error code (see error.hpp)
+     * @param   struct sockaddr* addr           [out]
      * @remarks
      *          ERROR_CONNECTION_CLOSED
      */
-    virtual return_t read(socket_t sock, tls_context_t* tls_handle, int mode, char* ptr_data, size_t size_data, size_t* cbread, struct sockaddr* addr = nullptr,
-                          socklen_t* addrlen = nullptr);
+    virtual return_t read(socket_t sock, tls_context_t* tls_handle, int mode, char* ptr_data, size_t size_data, size_t* cbread);
     /**
      * @brief   send
      * @param   socket_t        sock            [IN]