+++ sdk/net/basic/udp_server_socket.cpp
@@ -51,8 +51,8 @@ return_t udp_server_socket::close(socket_t sock, tls_context_t* tls_handle) {
     return ret;
 }
 
-return_t udp_server_socket::read(socket_t sock, tls_context_t* tls_handle, int mode, char* ptr_data, size_t size_data, size_t* size_read, struct sockaddr* addr,
-                                 socklen_t* addrlen) {
+return_t udp_server_socket::recvfrom(socket_t sock, tls_context_t* tls_handle, int mode, char* ptr_data, size_t size_data, size_t* size_read,
+                                     struct sockaddr* addr, socklen_t* addrlen) {
     return_t ret = errorcode_t::success;
     __try2 {
 #if 0
@@ -64,9 +64,9 @@ return_t udp_server_socket::read(socket_t sock, tls_context_t* tls_handle, int m
 #endif
 
 #if defined __linux__
-        int ret_recv = recvfrom(sock, ptr_data, size_data, 0, addr, addrlen);
+        int ret_recv = ::recvfrom(sock, ptr_data, size_data, 0, addr, addrlen);
 #elif defined _WIN32 || defined _WIN64
-        int ret_recv = recvfrom(sock, ptr_data, (int)size_data, 0, addr, addrlen);
+        int ret_recv = ::recvfrom(sock, ptr_data, (int)size_data, 0, addr, addrlen);
 #endif
         if (-1 == ret_recv) {
             ret = get_lasterror(ret_recv);
@@ -87,8 +87,8 @@ return_t udp_server_socket::read(socket_t sock, tls_context_t* tls_handle, int m
     return ret;
 }
 
-return_t udp_server_socket::send(socket_t sock, tls_context_t* tls_handle, const struct sockaddr* addr, socklen_t addrlen, const char* ptr_data,
-                                 size_t size_data, size_t* cbsent) {
+return_t udp_server_socket::sendto(socket_t sock, tls_context_t* tls_handle, const char* ptr_data, size_t size_data, size_t* cbsent,
+                                   const struct sockaddr* addr, socklen_t addrlen) {
     return_t ret = errorcode_t::success;
     __try2 {
 #if defined __linux__