+++ test/udpserver2/sample.cpp
@@ -33,30 +33,23 @@ t_shared_instance<t_cmdline_t<OPTION>> _cmdline;
 
 #define FILENAME_RUN _T (".run")
 
-ipaddr_acl acl;
-
-return_t accept_handler(socket_t socket, sockaddr_storage_t* client_addr, CALLBACK_CONTROL* control, void* parameter) {
-    return_t ret = errorcode_t::success;
-    bool result = false;
-
-    acl.determine(client_addr, result);
-    if (control) {
-        *control = result ? CONTINUE_CONTROL : STOP_CONTROL;
-    }
-    return ret;
-}
-
 return_t consume_routine(uint32 type, uint32 data_count, void* data_array[], CALLBACK_CONTROL* callback_control, void* user_context) {
     return_t ret = errorcode_t::success;
-    net_session_socket_t* network_session = (net_session_socket_t*)data_array[0];
-
-    char* buf = (char*)data_array[1];
+    net_session_socket_t* session_socket = (net_session_socket_t*)data_array[0];
+    byte_t* buf = (byte_t*)data_array[1];
     size_t bufsize = (size_t)data_array[2];
+    network_session* session = (network_session*)data_array[3];
     sockaddr_storage_t* addr = (sockaddr_storage_t*)data_array[5];
+
+    basic_stream bs;
+    std::string message;
+
     switch (type) {
-        case multiplexer_event_type_t::mux_dgram:
-            _logger->writeln("read %d msg [%.*s]", network_session->event_socket, (unsigned)bufsize, buf);
-            sendto((socket_t)network_session->event_socket, buf, bufsize, 0, (sockaddr*)addr, sizeof(sockaddr_storage_t));
+        case mux_dgram:
+            _logger->writeln("read %d msg [%.*s]", session_socket->event_socket, (unsigned)bufsize, buf);
+            // dump_memory (buf, bufsize, &bs, 16, 4);
+            // std::cout << bs << std::endl;
+            session->sendto((char*)buf, bufsize, addr);
             break;
     }
     return ret;
@@ -77,10 +70,6 @@ return_t echo_server(void* param) {
     fclose(fp);
 
     __try2 {
-        acl.add_rule("127.0.0.1", true);
-        acl.add_rule("::1", true);
-        acl.setmode(ipaddr_acl_t::whitelist);
-
         server_conf conf;
         conf.set(netserver_config_t::serverconf_concurrent_event, 1024)  // concurrent (linux epoll concerns, windows ignore)
             .set(netserver_config_t::serverconf_concurrent_tls_accept, 1)
@@ -89,9 +78,6 @@ return_t echo_server(void* param) {
 
         network_server.open(&handle_ipv4, AF_INET, port, &svr_sock, &conf, consume_routine, nullptr);
         network_server.open(&handle_ipv6, AF_INET6, port, &svr_sock, &conf, consume_routine, nullptr);
-
-        network_server.set_accept_control_handler(handle_ipv4, accept_handler);
-        network_server.set_accept_control_handler(handle_ipv6, accept_handler);
 
         network_server.consumer_loop_run(handle_ipv4, 2);
         network_server.consumer_loop_run(handle_ipv6, 2);